% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{add_left_sidebar}
\alias{add_left_sidebar}
\title{Add the left sidebar to an existing application.}
\usage{
add_left_sidebar(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Add the left sidebar to an existing application.
}
