% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{add_reset_button}
\alias{add_reset_button}
\title{Add the reset button to an existing application.}
\usage{
add_reset_button(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Add the reset button to an existing application.
}
