% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{add_right_sidebar}
\alias{add_right_sidebar}
\title{Add the right sidebar to an existing application.}
\usage{
add_right_sidebar(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Add the right sidebar to an existing application.
}
