% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{add_ui_menu_top}
\alias{add_ui_menu_top}
\title{Add UI Elements to the top menubar.}
\usage{
add_ui_menu_top(elementlist = NULL)
}
\arguments{
\item{elementlist}{list of UI elements to add to the top menu.}
}
\description{
This function registers UI elements at the top menubar.
}
\section{Shiny Usage}{

Call this function after creating elements in \code{program/ui_top_menu.R} to register
them to the application framework.
}

\examples{
require(shiny)

button <- actionButton("inputId", "Label")
 
add_ui_menu_top(list(button))

}
\seealso{
\link[periscope]{add_ui_sidebar_basic}

\link[periscope]{add_ui_body}
}
