% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile_AvailableTypes}
\alias{downloadFile_AvailableTypes}
\title{downloadFile Helper}
\usage{
downloadFile_AvailableTypes()
}
\value{
a vector of all supported types
}
\description{
Returns a list of all supported types
}
\seealso{
\link[periscope]{downloadFileButton}

\link[periscope]{downloadFile}
}
