% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFile.R
\name{downloadFile_ValidateTypes}
\alias{downloadFile_ValidateTypes}
\title{downloadFile Helper}
\usage{
downloadFile_ValidateTypes(types)
}
\arguments{
\item{types}{list of types to test}
}
\value{
the list input given in types
}
\description{
Checks a given list of file types and warns if an invalid type is included
}
\section{Example}{

\code{downloadFile_ValidateTypes(c("csv", "tsv"))}
}

\seealso{
\link[periscope]{downloadFileButton}

\link[periscope]{downloadFile}
}
