% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{get_url_parameters}
\alias{get_url_parameters}
\title{Get URL Parameters}
\usage{
get_url_parameters(session)
}
\arguments{
\item{session}{shiny session object}
}
\value{
named list of url parameters and values.  List may be empty if
no URL parameters were passed when the application instance was launched.
}
\description{
This function returns any url parameters passed to the application as
a named list.  Keep in mind url parameters are always user-session scoped
}
