% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periscope-package.R
\name{has_ffmpeg}
\alias{has_ffmpeg}
\title{Helper function to check if ffmpeg is available}
\usage{
has_ffmpeg()
}
\value{
Returns \code{TRUE} if ffmpeg is on the PATH.
}
\description{
Helper function to check if ffmpeg is available
}
