% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{remove_left_sidebar}
\alias{remove_left_sidebar}
\title{Removes the left sidebar from an existing application.}
\usage{
remove_left_sidebar(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Removes the left sidebar from an existing application.
}
