% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_template.R
\name{remove_reset_button}
\alias{remove_reset_button}
\title{Remove the reset button from an existing application.}
\usage{
remove_reset_button(location)
}
\arguments{
\item{location}{path of the existing application.}
}
\description{
Remove the reset button from an existing application.
}
