% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{ui_tooltip}
\alias{ui_tooltip}
\title{Insert a standardized tooltip}
\usage{
ui_tooltip(id, label = "", text = "")
}
\arguments{
\item{id}{character id for the tooltip object}

\item{label}{text label to appear to the left of the tooltip image}

\item{text}{tooltip text shown when the user hovers over the image}
}
\description{
This function inserts a standardized tooltip image, label (optional),
and hovertext into the application UI
}
